+++ packages/create-mechanic/function-examples/instagram-story-generator/index.js
@@ -19,19 +19,18 @@ export const handler = ({ inputs, done, frame, useDrawLoop }) => {
 
   // stuff needed for the looping
   const isPlaying = useRef(true);
-  const frameCount = useDrawLoop(isPlaying.current);
+  const { timestamp } = useDrawLoop(isPlaying.current);
   const lines = text.split(" ");
-  const durationInFrames = duration * 60;
 
   // function to determine when to end the animation
   useEffect(() => {
-    if (durationInFrames > frameCount) {
+    if (duration > timestamp) {
       frame();
     } else if (isPlaying.current) {
       isPlaying.current = false;
       done();
     }
-  }, [frameCount]);
+  }, [timestamp]);
 
   // colors and font sizes
   const textColor = backgroundColor;
@@ -44,7 +43,7 @@ export const handler = ({ inputs, done, frame, useDrawLoop }) => {
   // this is an array where will  store the circles
   const circles = [];
 
-  for (let i = 0; i < Math.min(Math.floor(frameCount / 15), 20); i++) {
+  for (let i = 0; i < Math.min(Math.floor(timestamp * 4), 20); i++) {
     circles.push(
       <Circle
         key={i}